--- a/cbor_gen.go
+++ b/cbor_gen.go
@@ -11,21 +11,20 @@ import (
 	xerrors "golang.org/x/xerrors"
 )
 
-// NOTE: This is a generated file, but it has been modified to encode the
-// bitfield big.Int as a byte array. The bitfield is only a big.Int because
-// thats a convenient type for the operations we need to perform on it, but it
-// is fundamentally an array of bytes (bits)
-
 var _ = xerrors.Errorf
+
+var lengthBufNode = []byte{130}
 
 func (t *Node) MarshalCBOR(w io.Writer) error {
 	if t == nil {
 		_, err := w.Write(cbg.CborNull)
 		return err
 	}
-	if _, err := w.Write([]byte{130}); err != nil {
+	if _, err := w.Write(lengthBufNode); err != nil {
 		return err
 	}
+
+	scratch := make([]byte, 9)
 
 	// t.Bitfield (big.Int) (struct)
 	{
@@ -47,7 +46,7 @@ func (t *Node) MarshalCBOR(w io.Writer) error {
 		return xerrors.Errorf("Slice value in field t.Pointers was too long")
 	}
 
-	if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajArray, uint64(len(t.Pointers)))); err != nil {
+	if err := cbg.WriteMajorTypeHeaderBuf(scratch, w, cbg.MajArray, uint64(len(t.Pointers))); err != nil {
 		return err
 	}
 	for _, v := range t.Pointers {
@@ -60,8 +59,9 @@ func (t *Node) MarshalCBOR(w io.Writer) error {
 
 func (t *Node) UnmarshalCBOR(r io.Reader) error {
 	br := cbg.GetPeeker(r)
+	scratch := make([]byte, 8)
 
-	maj, extra, err := cbg.CborReadHeader(br)
+	maj, extra, err := cbg.CborReadHeaderBuf(br, scratch)
 	if err != nil {
 		return err
 	}
@@ -75,7 +75,7 @@ func (t *Node) UnmarshalCBOR(r io.Reader) error {
 
 	// t.Bitfield (big.Int) (struct)
 
-	maj, extra, err = cbg.CborReadHeader(br)
+	maj, extra, err = cbg.CborReadHeaderBuf(br, scratch)
 	if err != nil {
 		return err
 	}
@@ -99,7 +99,7 @@ func (t *Node) UnmarshalCBOR(r io.Reader) error {
 	}
 	// t.Pointers ([]*hamt.Pointer) (slice)
 
-	maj, extra, err = cbg.CborReadHeader(br)
+	maj, extra, err = cbg.CborReadHeaderBuf(br, scratch)
 	if err != nil {
 		return err
 	}
@@ -129,23 +129,28 @@ func (t *Node) UnmarshalCBOR(r io.Reader) error {
 	return nil
 }
 
+var lengthBufKV = []byte{130}
+
 func (t *KV) MarshalCBOR(w io.Writer) error {
 	if t == nil {
 		_, err := w.Write(cbg.CborNull)
 		return err
 	}
-	if _, err := w.Write([]byte{130}); err != nil {
+	if _, err := w.Write(lengthBufKV); err != nil {
 		return err
 	}
+
+	scratch := make([]byte, 9)
 
 	// t.Key ([]uint8) (slice)
 	if len(t.Key) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.Key was too long")
 	}
 
-	if _, err := w.Write(cbg.CborEncodeMajorType(cbg.MajByteString, uint64(len(t.Key)))); err != nil {
+	if err := cbg.WriteMajorTypeHeaderBuf(scratch, w, cbg.MajByteString, uint64(len(t.Key))); err != nil {
 		return err
 	}
+
 	if _, err := w.Write(t.Key); err != nil {
 		return err
 	}
@@ -159,8 +164,9 @@ func (t *KV) MarshalCBOR(w io.Writer) error {
 
 func (t *KV) UnmarshalCBOR(r io.Reader) error {
 	br := cbg.GetPeeker(r)
+	scratch := make([]byte, 8)
 
-	maj, extra, err := cbg.CborReadHeader(br)
+	maj, extra, err := cbg.CborReadHeaderBuf(br, scratch)
 	if err != nil {
 		return err
 	}
@@ -174,7 +180,7 @@ func (t *KV) UnmarshalCBOR(r io.Reader) error {
 
 	// t.Key ([]uint8) (slice)
 
-	maj, extra, err = cbg.CborReadHeader(br)
+	maj, extra, err = cbg.CborReadHeaderBuf(br, scratch)
 	if err != nil {
 		return err
 	}