--- a/go.mod
+++ b/go.mod
@@ -2,11 +2,11 @@ module github.com/ipfs/go-hamt-ipld
 
 require (
 	github.com/ipfs/go-block-format v0.0.2
-	github.com/ipfs/go-cid v0.0.4
+	github.com/ipfs/go-cid v0.0.6-0.20200501230655-7c82f3b81c00
 	github.com/ipfs/go-ipld-cbor v0.0.4
 	github.com/ipfs/go-ipld-format v0.0.2 // indirect
 	github.com/spaolacci/murmur3 v1.1.0
-	github.com/whyrusleeping/cbor-gen v0.0.0-20200501014322-5f9941ef88e0
+	github.com/whyrusleeping/cbor-gen v0.0.0-20200504204219-64967432584d
 	golang.org/x/sys v0.0.0-20190626221950-04f50cda93cb // indirect
 	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543
 )