Some noticeable improvements here, not as significant as other places, but still pretty decent:

Old:
```
oos: linux
goarch: amd64
pkg: github.com/ipfs/go-hamt-ipld
BenchmarkSerializeNode-12    	   30603	     34178 ns/op	   16736 B/op	     302 allocs/op
BenchmarkSet/1000/5-12       	     319	   3480615 ns/op
BenchmarkSet/1000/8-12       	     468	   3698476 ns/op
BenchmarkSet/10000/5-12      	      34	  33712720 ns/op
BenchmarkSet/10000/8-12      	      50	  35598423 ns/op
BenchmarkSet/100000/5-12     	       3	 341730111 ns/op
BenchmarkSet/100000/8-12     	       3	 363837179 ns/op
BenchmarkFind/find-10k-12    	    4221	    282219 ns/op	   50345 B/op	    3211 allocs/op
BenchmarkFind/find-100k-12   	    3160	    408864 ns/op	  108236 B/op	    7018 allocs/op
BenchmarkFind/find-1m-12     	    4224	    300982 ns/op	   86031 B/op	    5415 allocs/op
BenchmarkFind/find-10k-bitwidth-5-12         	   17031	     69950 ns/op	   13449 B/op	     851 allocs/op
BenchmarkFind/find-100k-bitwidth-5-12        	   15105	     80791 ns/op	   21461 B/op	    1363 allocs/op
BenchmarkFind/find-1m-bitwidth-5-12          	   14520	     80031 ns/op	   22368 B/op	    1426 allocs/op
PASS
ok  	github.com/ipfs/go-hamt-ipld	51.436s
```

New:
```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-hamt-ipld
BenchmarkSerializeNode-12    	   43467	     30901 ns/op	   14480 B/op	     251 allocs/op
BenchmarkSet/1000/5-12       	     332	   3372082 ns/op
BenchmarkSet/1000/8-12       	     452	   3613512 ns/op
BenchmarkSet/10000/5-12      	      34	  32244894 ns/op
BenchmarkSet/10000/8-12      	      57	  33884408 ns/op
BenchmarkSet/100000/5-12     	       3	 342868418 ns/op
BenchmarkSet/100000/8-12     	       3	 364767968 ns/op
BenchmarkFind/find-10k-12    	    4038	    279104 ns/op	   50172 B/op	    3121 allocs/op
BenchmarkFind/find-100k-12   	    3393	    378689 ns/op	  107780 B/op	    6373 allocs/op
BenchmarkFind/find-1m-12     	    3540	    296424 ns/op	   85957 B/op	    5372 allocs/op
BenchmarkFind/find-10k-bitwidth-5-12         	   17947	     67492 ns/op	   13280 B/op	     812 allocs/op
BenchmarkFind/find-100k-bitwidth-5-12        	   15703	     75456 ns/op	   21364 B/op	    1263 allocs/op
BenchmarkFind/find-1m-bitwidth-5-12          	   15655	     76398 ns/op	   22300 B/op	    1355 allocs/op
PASS
ok  	github.com/ipfs/go-hamt-ipld	50.790s
```