@whyrusleeping 2020-05-04
Mostly the reason for the benchmarks not looking that great is that the benchmarks operate over an unflushed tree, meaning everything ends up in memory pretty quickly